Adopt file layout of ruby/ruby
The current layout is hard to sync with ruby/ruby
